rename proposals.iou to proposals.metrics
