+++ proposals/__init__.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
 
-from .iou import compute_iou
+from .metrics import *
 from .super_vertex import super_vertex
 from .stat_cluster import get_majority
 from .generate_basic_proposals import (filter_clusters, save_proposals,
